pyproject.toml
[tool.poetry]
name = "airflow-k8s-operator"
version = "0.1.0"
description = ""
authors = ["javier162380 <javierllorente16@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
apache-airflow = {version = "1.10.9", extras = ["kubernetes", "gcp"]}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
